VideoLocalizing/requirements.txt
absl-py==0.9.0
altgraph==0.17
appdirs==1.4.3
astor==0.7.1
astroid==2.4.1
Augmentor==0.2.8
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.4.1
colorama==0.4.3
colour==0.1.5
cycler==0.10.0
dask==2.15.0
decorator==4.4.2
dominate==2.5.1
future==0.18.2
gast==0.2.2
google-api-core==1.17.0
google-auth==1.15.0
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
googletrans==2.4.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib==1.0.4
imutils==0.5.3
isort==4.3.21
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lmdb==0.98
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.4
numpy==1.16.1
olefile==0.46
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
opt-einsum==3.2.1
pefile==2019.4.18
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycuda==2019.1.2
pygame==1.9.6
PyInstaller==3.6
pylint==2.5.2
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
python-dateutil==2.8.1
pytools==2020.1
pytz==2020.1
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
requests==2.23.0
rsa==4.0
scikit-image==0.14.2
scipy==1.1.0
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
toml==0.10.1
toolz==0.10.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.2.2.post3
tqdm==4.45.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
